+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix response to requesting deleted or private datasets: this now results in a 404 response, not a 500.
 - Set env variables for database and Solr index in scripts for running unit tests locally.
+- Move `_is_true()` to helpers.
 - Add required `author_uri` to test data.
 
 ## [0.1.12](https://github.com/berlinonline/ckanext-datasetsnippets/releases/tag/0.1.12)